+++ runtime/include/icpp.hpp
@@ -177,5 +177,7 @@ private:
 
 ICPP_API void print(std::uint64_t val);
 ICPP_API void print_hex(std::uint64_t val);
+ICPP_API void set_env(std::string_view key, std::string_view value,
+                      bool overwrite = true);
 
 } // namespace icpp